move: card 컴포넌트 위치 이동/fix: postId 페이지 아래 여백 추가
