+++ src/components/DivToButton.jsx
@@ -1,7 +1,7 @@
 //DivToButton.jsx
 //div 태그로 button같은 효과를 얻기 위한 컴포넌트입니다.
 
-const DivToButton = ({ children, onClick, Ref, className }) => {
+const DivToButton = ({ children, onClick, Ref, className, disabled }) => {
   const handleKeyDown = function (e) {
     if (e.code === "Enter" || e.code === " ") {
       e.preventDefault();
@@ -17,6 +17,7 @@ const DivToButton = ({ children, onClick, Ref, className }) => {
       onClick={onClick}
       ref={Ref}
       className={className}
+      aria-disabled={disabled}
     >
       {children}
     </div>
