+++ src/components/Header.module.scss
@@ -32,7 +32,7 @@
   display: flex;
   align-items: center;
   color: #4a494f;
-  font-family: Poppins;
+  font-family: Poppins, sans-serif;
   font-size: 19.971px;
   line-height: normal;
   &::before {
