+++ src/components/PostId/Card/CardFrom.module.scss
@@ -1,4 +1,4 @@
-@import "../../global.scss";
+@import "../../../global.scss";
 
 .frame {
   display: flex;
