+++ src/components/PostId/Card/CreatedDay.module.scss
@@ -1,4 +1,4 @@
-@import "../../global.scss";
+@import "../../../global.scss";
 
 .time {
   color: $gray4;
